Code/shash_code/requirements.txt
Pillow==10.4.0
pdf2image==1.17.0
